Commit after second feedback

+++ src/fr/sciforma/apietnic/business/provider/OrganizationFieldProvider.java
@@ -20,7 +20,7 @@ public class OrganizationFieldProvider implements FieldProvider<Organization> {
     public void init() {
         fields = new ArrayList<>();
         fields.add(SciformaField.builder().name("Description").type(FieldType.STRING).build());
-        fields.add(SciformaField.builder().name("Internal ID").type(FieldType.DECIMAL).build());
+        fields.add(SciformaField.builder().name("Internal ID").type(FieldType.INTEGER).build());
         fields.add(SciformaField.builder().name("Manager 1").type(FieldType.USER).build());
         fields.add(SciformaField.builder().name("Manager 2").type(FieldType.USER).build());
         fields.add(SciformaField.builder().name("Manager 3").type(FieldType.USER).build());
